finishedalmost

+++ getto100/src/App.jsx
@@ -8,20 +8,39 @@ function App() {
   const [startButton, setStartButton] = useState(false);
 
   function addGamer(userName) {
-    const newPlayer = {
-      userName: userName,
-      scores: [],
-      initialNumber: Math.floor(Math.random() * 100),
-      resetCounter: 0,
-    };
-    const updatedPlayers = [...players, newPlayer];
-    setPlayers(updatedPlayers);
     const savedPlayers = JSON.parse(localStorage.getItem("players") || "[]");
-    localStorage.setItem(
-      "players",
-      JSON.stringify([...savedPlayers, newPlayer])
+
+    const existingPlayer = savedPlayers.find(
+      (player) => player.userName === userName
     );
+
+    let newPlayerList;
+
+    if (existingPlayer) {
+      const alreadyIn = players.some(
+        (p) => p.userName === existingPlayer.userName
+      );
+      if (!alreadyIn) {
+        newPlayerList = [...players, existingPlayer];
+        setPlayers(newPlayerList);
+      }
+    } else {
+      const newPlayer = {
+        userName,
+        scores: [],
+        initialNumber: Math.floor(Math.random() * 100),
+        resetCounter: 0,
+      };
+      newPlayerList = [...players, newPlayer];
+      setPlayers(newPlayerList);
+
+      localStorage.setItem(
+        "players",
+        JSON.stringify([...savedPlayers, newPlayer])
+      );
+    }
   }
+
   return (
     <>
       <Header
